fix(mypy): resolve 66 pre-existing type errors in cli.* (Phase 3 pre-work)
Targeted minimal fixes across 28 files to make mypy src/backlink_publisher/cli/
report zero errors, enabling [mypy-backlink_publisher.cli.*] disallow_untyped_defs
in a follow-on commit.

Fix categories:
- cast() for "Returning Any" (pipeline_orchestrator, phase0_seal, weights,
  _footprint_baseline, canary_targets, canary_seed, _payload.py)
- type: ignore[unreachable] for intentional dead-code guards after NoReturn calls
  (_plan_check_schema, medium recipe, plan_backlinks core/_work_themed, _resume,
  canary_seed)
- loop-variable rename to avoid mypy's dict/str confusion in _seal_init and
  phase0_seal (_body → _body_str, _new_body → _new_body_str)
- Literal cast for DedupStore.transition state arg in _dedup_gate
- str | None → str narrowing with "or ''" for emit_error/emit_envelope_and_exit
  in _publish_helpers
- Path→str conversion for _check_all() in health_check
- getattr() instead of .get() on Config object in pr_opportunities
- cast(dict[str, Any], ...) for _TEMPLATES values in _payload.py
- cast(StorageState→dict) for context.storage_state() in _driver_impl
- Missing return after handle_error() in comment.py
- mode or "draft" fallback for create_checkpoint in publish_backlinks._engine
- Renamed loop variable in plan_backlinks._engine to fix no-redef error
- cast(list[dict], ...) for SprayCandidate.row comprehensions; avoids CC increase
- __main__.py trio: main(); raise SystemExit(0) instead of main() or 0

+++ src/backlink_publisher/cli/_resume.py
@@ -259,8 +259,8 @@ def _publish_one_resume_item(
     banner_emit: Any,
     run_id: str,
     args: Any,
-    throttle_min: float,
-    throttle_max: float,
+    throttle_min: int,
+    throttle_max: int,
 ) -> None:
     """Phase 6 body: process exactly one resume item.
 
@@ -360,7 +360,7 @@ def _publish_one_resume_item(
         )
         # Surface the real class (AuthExpiredError), not exit-3's "DependencyError".
         emit_error(str(exc), exit_code=3, error_class=type(exc).__name__)
-        return
+        return  # type: ignore[unreachable]
     except BannerUploadError as exc:
         _record_resume_failure(
             run_id, item, exc,
@@ -370,7 +370,7 @@ def _publish_one_resume_item(
     except DependencyError as exc:
         record_failure(row, platform, error_class="dependency", run_id=run_id)
         emit_error(str(exc), exit_code=3)
-        return
+        return  # type: ignore[unreachable]
     except ExternalServiceError as exc:
         _record_resume_failure(
             run_id, item, exc,
@@ -530,7 +530,7 @@ def _run_resume(args: Any) -> None:
         ckpt = checkpoint.load_checkpoint(run_id)
     except (ValueError, FileNotFoundError) as exc:
         emit_error(str(exc), exit_code=2)
-        return
+        return  # type: ignore[unreachable]
 
     config = load_config()
     banner_emit = _make_banner_emit()
